Introduce uniform method of handling encoding
Under gssapi._utils, there are now two new methods:
one for use inside the library, and one for use
outside.

For use from the outside, we have `set_encoding`, which
sets the encoding to be used for normal strings.  By
default, it is UTF-8.

From the inside, we have '_get_encoding', which gets the
current encoding.