Detect support for extensions
Previously, setup.py attempted to compile all of the modules for
the various extensions for GSSAPI.   Now, setup.py will attempt
to load the GSSAPI library (specified by the GSSAPI_MAIN_LIB
environment variable, defaulting to 'libgssapi_krb5.so') to look
for a 'canary' function to detect support for the extension.
If the canary function is not found, the extension is not compiled.
This functionality can be disabled by setting the GSSAPI_SUPPORT_DETECT
environment variable to 'false'.

Additionally, the GSSAPI_MECHS environment variable may be set to
control with mechansim-specific files are compiled.  It defaults to
'krb5'.

Finally, the GSSAPI_LINKER_ARGS and GSSAPI_COMPILER_ARGS environment
variables may be set to control the compiler and linker arguments.
They default to the output of the 'krb5-config gssapi' command
(with either the '--libs' or '--cflags', respectively).