Introduce Channel Bindings Support
This commit introduces support for channel bindings.
The `ChannelBindings` class, found in `raw`, may be used
to create a set or channel bindings that may be passed to
both the high- and low-level APIs.  Additionally, a new
`IntEnum` named `AddressType` has been introduced; `AddressType`
contains potential values for the initiator and acceptor address-type
fields in the `ChannelBindings` class.