Note that address_type of None is AF_UNSPEC
This commit adds a note indicating that setting addrtype to 0
(as done by the calloc statement above) in the channel bindings
struct is equivalent to setting it to GSS_C_AF_UNSPEC
(i.e. what was AddressType.unspecified).

Additionally, it uncomments the `null` member of AddressType
(which is *not* equivalent to `0`) and comments out the
`unspecified` member.