Clean Up Setup.py Before Publish
`setup.py` will now automatically detect support for Cython and
compile with Cython when available.  Otherwise, it will try to
compile with the C files.  `MANIFEST.in` now keeps the C files
along with the Cython files, so that running `setup.py sdist`
will keep the C files in the generated source distribution.

Finally, the "author" of the project has been changed to
"The Python GSSAPI Team".  The license has been updated to
reflect this.  The actual license itself has been changed into
a more streamlined ISC license.