Don't compile from cython when installing sdist
It is reccomended not to have setup.py build from the Cython files
when users are installing from an sdist, even if they have Cython
installed.  Thus, the sdist command now creates a temorary file
included in the sdist file that instructs setup.py not to use Cython.
If setup.py detects this file, it will not even attempt to look for
Cython; it simply compiles from C.