Skip E402 on Cython Files, Fix E402 Errors
pep8 1.6.0 introduced E402, which warns about module-level imports
not being at the top of the file.  However, due to a quirk in Cython,
we need to set a constant at the top of the files, so disable this
check for now.

This also fixes an E402 error in `creds.py`.