Low-Level: Document Specific Errors
Now that we have specific errors (found in `gssapi.raw.exceptions`),
we should document which errors can be thrown.

This commit does so for the low-level API.  In the case where no
errors were documented, the only error documented was the generic
GSSError, otherwise, GSSError was removed, and the specific listed
errors were added (except in the case where GSS_S_FAILURE is
specifically listed, in which case so is GSSError).

Note that non-RFC extensions did not have specific errors documented,
since they do not have a standardized list of specific errors.

Part of #11