Implement RFC 6680 (Low-Level)
This commit introduces optional support for RFC 6680 to
the low-level API.  The `get_name_attribute` and
`set_name_attribute` methods adhere to the pseudo-API
(which uses a single call to get/set multiple values
on a single attribute) instead of the C bindings
(which require the use of a state variable to get
multiple values, and require multiple calls to
add multiple values to a single attribute).

Note that gss_display_name_ext is not tested, since
it is not implemented by MIT krb5.  Additionally,
in order to test the get/set/delete attribute methods,
you will have to install the demo greet plugin that
comes with krb5.  Otherwise, the tests will be skipped.

Part of #4

Also-Authored-By: Simo Sorce <simo@redhat.com>