Implement RFC 6680 (High-Level)
This commit introduces optional support for RFC 6680 to
the high-level API.

For attribute access, a new property named `attributes` was
introduced to the Name class.  This presents a `MutableMapping`
interface to the Name's attributes.  When iterables are assigned
to attributes (not including strings and bytes), they are considered
to be multiple values to be assigned to the attribute.  Additionally,
attribute names (but not values) are automatically encoded if they
are in text (and not bytes) form

For inquiry, appropriate properties were added to the `Name` class
(`is_mech_name` and `mech`).  `display_as` may be used to call
`display_name_ext`, and a `composite` argument was introduced to
both the `export` method and the constructor.

Closes #4