Actually use acceptor creds when passed in
Previously, the condition on acceptor credentials was inverted,
such that when `acceptor_creds` was `None`, it would attempt to use
the value of the underlying raw `gss_cred_id_t`, and when it was
not `None`, the actual credentials passed in would be
`GSS_C_NO_CREDENTIAL`.  This has been corrected.

Additionally, two new tests have been added.  One checks that passing
None for acceptor credentials is ok, and the second properly tests that
`accept_sec_context` received S4U2Proxy delegated credentials in the
appropriate circumstances.