Introduce util for locating the plugins dir
This commit adds a method to `gssapi.tests._utils`
to locate the `krb5` plugin directory.

The logic works roughly as such:

If there is a LD_LIBRARY_PATH set, start searching there.
First, assume that LD_LIBRARY_PATH refers to a normal
installation root (like '/usr/lib'), and so search for
a 'krb5/plugins' directory contain .so files somewhere
amongst its subdirectories.

If that doesn't work, assume LD_LIBRARY_PATH points right
into a krb5 source directory's lib folder, so search for a
'plugins' directory which contains .so files,
in the parent directory of LD_LIBRARY_PATH.

If no LD_LIBRARY_PATH is set, or the above fails, use the
'$(krb5-config --prefix)/lib' instead of LD_LIBRARY_PATH
in the first case.

If multiple results are found, use the shortest path.

Also-Authored-By: Robbie Harwood (frozencemetery) <rharwood@redhat.com>