Introduce the gssapi-console utilitiy
This utility provides an interactive Python session that automatically
sets up and tears down a self-contained krb5 realm.  It does this
through the use of the `gssapi.tests.k5test` module.  The created
`K5Realm` class is accessible inside the session as `REALM`, provides
methods for manipulating the realm setup (such as creating principals,
kinit-ing, etc).  Additionally, `gssapi` and `gssapi.raw as gb` are
automatically imported.

Just like the normal Python interpreter, when no file is passed,
interactive mode is used.  Otherwise if a file is passed, the file
is simply run in the context of the realm, unless the `-i` option
is used.  Comma-separated boolean options can also be passed to
the realm constructor via the `--realm-args` option, in the form
of `key=(true|false)`.