Fix multiple -DHAS_GSSAPI_EXT_H compile flags
Previously, '-DHAS_GSSAPI_EXT_H' was printed multiple times for each
extension, since the list of compile flags for each extension was
actually just the global list.  To correct this, we add
'-DHAS_GSSAPI_EXT_H' once to the global list, instead of trying
to add it to each extension's local list.