High-Level Documentation Updates
This commit updates the high-level documentation.  Typos were
corrected, improvements were made for readability, and more
links were inserted.

Additionally, two custom Sphinx extensions were added to help
with the process:

The first extension catches unknown class references and attempts
to track them down.  The class is searched for as if it were an
`:any:` reference.  The results from this searched are used, prefering
results in the current API level (high-level vs low-level).

The second extension adds a convinience role, `:requires-ext`, which
inserts a link and some text indicating that a particular extension
is required.  If the extension is an RFC, a link to the RFC is
generated, otherwise, a link to the low-level API module is generated.
In the latter case, the text of the link is generated using the first
line of the docstring of the low-level module, if it is present.
Otherwise, the partial name of the module is used (the '%s' part
of 'ext_%s')

Finally, as per above, the docstrings of several extensions were
updated.