Docs: ToC Tweaks
This commit bumps the docs theme version to get support for deeper
nesting on the nav bar.  It increases the ToC depth to 3, and hides
the ToC from the intro page to avoid redundancy.

New headings were added to the high-level API docs to take advantage of
the increased ToC depth.