Ensure that delegated_creds is non-null when set
Previously, in the high-level API, we would unconditionally set
wrap the return value of delegated_creds, which could result in
delegated_creds being set to the default credentials when no
delegated creds were returned (since the call would be
`Credentials(None)`).  This changes the logic so that delegated_creds
stays set as `None` unless delegated_creds are actually returned.

Partial solution for #96