Force overwrite on cred_store tests
Certain versions of krb5 (>= 1.14 or anything that backported the
patch, including RHEL/Centos/Fedora) object (probably rightly so) to
duplicate credentials already existing in the ccache.  Tell those
versions to sit down and shut up.

This is the third and final commit in a series to enable running our
test suite on RHEL-7.