The http service name is HTTP
Service names are generally case sensitive in Kerberos implementations (with
the notable exception of Microsoft Active Directory that treats them in a
case-insensitive way). For unknown reasons the service name used for the
http protocol has been historically uppercased. So the examples should use
HTTP otherwise implementors may use http and fail to interoperate with
servers that follow the proper case sensitivity rules for Kerberos principal
names as defined in RFC4120.

Signed-off-by: Simo Sorce <simo@redhat.com>