Make requires_rfc.py properly set parents of nodes
The requires_rfc.py docs extension was not properly setting the parent
node when adding child text nodes.  This caused a hiccup in newer
versions of Sphinx, meaning we could not generate new documentation.

This commit should fix that by using methods to clear and set the
children, as opposed to directly setting the child list.