Add channel binding extension support
This adds support for the channel binding extension
from the Kitten WG. This adds two new calls to the
raw api: create_sec_context and set_context_flags,
allowing ret_flags_understood to be set for mechs
that support it. This flag is exposed as the
channel_bound requirement flag.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>