OIDs: Copy constructor take ownership
When constructing via a the cpy parameter, cpy
would maintain ownership; if cpy._free_on_dealloc
was set to true, the new object could have memory
freed while still maintaining a reference to it.
This fixes the issue by having the new object take
ownership. To perform a memory copy, use the elements
argument to the constructor.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>