Travis build stages
This fixes our Travis auto-deployment mechnanisms, and switches
over to making use of Travis build stages along the way.

There are now three stages:

1. verify: this *just* runs the verification suite, which currently
   consists of the flake8 tests.  The verification is run under
   both Python 2 and Python 3, using the fedora:rawhide image

2. test: this runs all of non-verification tests, under the full
   test matrix we were using before

3. deploy: this runs our deployements, as before

The deploy phase now precompiles the sdist, and asks for that to be
uploaded instead of regenerating the code before upload.  This is a bit
of a hack.

Finally, the actual Travis scripts themselves were refactored into
a series of smaller files and utility functions, which makes the code
a bit more readable in some cases, and allows individual functions
to be called separately (useful in testing stuff out, and in creating
new build stages).