Implement support for GSSAPI extension RFC 4178
RFC 4178 provides two support API calls that enable the caller to
manipulate the set of acceptable security mechanisms used in SPNEGO
protocol; for the given credentials, the gss_get_neg_mechs call is
used to indicate the current set of security mechanisms available
for negotiation, and the gss_set_neg_mechs call is used to specify
the set of security mechanisms avaiable for negotiation. Since
gss_get_neg_mechs is not implemented by MIT krb5, we are only
implementing the raw interface for the latter call. Note that
although RFC 4178 did not specify that the mech_set argument cannot
be an empty set, we are forcing it to be non empty in the low-level
API here since passing an empty set will always trigger an error in
MIT krb5 implementation.