Change sdist and build to use PEP517
Changes the sdist to no longer include the cythonised .c files and rely
on the PEP517 features to define Cython as a build requirement. This
ensures that when building this package from the sdist, the process will
include any bugfixes for Cython that have been fixed after the
python-gssapi version has been released.

The removal of the .c files also simplifies the setup.py by removing
more functionality that was tied specifically to setuptools or even
distutils making it more flexible for the future.

The CI process has been updated to test against this new artifact rather
than testing against an build_ext --inplace instance to better replicate
how end users will be using the library.

Signed-off-by: Jordan Borean <jborean93@gmail.com>