Bump version to 1.8.1 with Python 3.7 requires
This update bumps the `python_requires` to be `>=3.7` as the sdist build
requires PEP 517 support that is somewhat only guaranteed with the pip
that was shipped with Python 3.7 or greater. Since 3.6 has been end of
life since December 2021 and the remaining Linux distributions that
still use it have system provided packages this shouldn't be a surprise.

Also use CentOS 9 Stream in the CI tests to test this on an actual
RHEL-ish host on the version of Python actually supported.

Signed-off-by: Jordan Borean <jborean93@gmail.com>